+++ anubis-management-api/anubis/policies/routers.py
@@ -1,12 +1,12 @@
 from typing import List, Optional
 from fastapi import Depends, APIRouter, HTTPException, status, Response, Header
 from . import operations, models, schemas
-from dependencies import get_db
+from ..dependencies import get_db
 from sqlalchemy.orm import Session
-from tenants import operations as so
-from wac import serialize as w_serialize
-from rego import serialize as r_serialize
-import default
+from ..tenants import operations as so
+from ..wac import serialize as w_serialize
+from ..rego import serialize as r_serialize
+import anubis.default as default
 import jwt