+++ RELEASE_NOTES.md
@@ -9,6 +9,7 @@
 ### Bug fixes
 
 - Revised method mappings, acl:Control behaviour
+- Fix readme
 
 ### Documentation