+++ RELEASE_NOTES.md
@@ -18,6 +18,8 @@
 
 ### Documentation
 
+- Improve documentation and provide ttl version of the vocabulary.
+
 ### Continuous Integration
 
 ### Technical debt